move server actions into utils folder
